simple-gamify/requirements.txt
apispec==6.0.2
click==8.1.3
debugpy==1.6.6
dnspython==2.3.0
Flask==2.2.2
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2==2.9.5
python-dotenv==0.21.1
SQLAlchemy==2.0.3
SQLAlchemy-Utils==0.40.0
typing_extensions==4.4.0
webargs==8.2.0
Werkzeug==2.2.2
pyyaml
jsonschema